Move RequestOptions into separate file
