+++ apps/dh/api-dh/source/DataHub.WebApi/Modules/ProcessManager/Requests/RequestOperations.cs
@@ -12,16 +12,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using System.Security.Claims;
-using Energinet.DataHub.ProcessManager.Orchestrations.Abstractions.Processes.BRS_023_027.V1.Model;
 using Energinet.DataHub.ProcessManager.Orchestrations.Abstractions.Processes.Shared.BRS_026_028;
 using Energinet.DataHub.WebApi.Clients.MarketParticipant.v1;
 using Energinet.DataHub.WebApi.Extensions;
-using Energinet.DataHub.WebApi.Modules.Common.Types;
 using Energinet.DataHub.WebApi.Modules.ProcessManager.Requests.Client;
 using Energinet.DataHub.WebApi.Modules.ProcessManager.Requests.Extensions;
 using Energinet.DataHub.WebApi.Modules.ProcessManager.Requests.Types;
-using MeteringPointType = Energinet.DataHub.Edi.B2CWebApp.Clients.v1.MeteringPointType;
 
 namespace Energinet.DataHub.WebApi.Modules.ProcessManager.Requests;
 
@@ -46,50 +42,6 @@ public static class RequestOperations
 
         var selectedActor = await marketParticipantClient.ActorGetAsync(associatedActor);
         return new RequestOptions(user, selectedActor.MarketRole.EicFunction);
-    }
-
-    public class RequestOptions(ClaimsPrincipal user, EicFunction marketRole)
-    {
-        public IEnumerable<Option<CalculationType>> GetCalculationTypes()
-        {
-            var calculationTypes = new List<Option<CalculationType>>();
-
-            if (user.HasRole("request-aggregated-measured-data:view"))
-            {
-                calculationTypes.Add(new Option<CalculationType>(CalculationType.Aggregation));
-                calculationTypes.Add(new Option<CalculationType>(CalculationType.BalanceFixing));
-            }
-
-            if (user.HasRole("request-wholesale-settlement:view"))
-            {
-                calculationTypes.Add(new Option<CalculationType>(CalculationType.WholesaleFixing));
-                calculationTypes.Add(new Option<CalculationType>(CalculationType.FirstCorrectionSettlement));
-                calculationTypes.Add(new Option<CalculationType>(CalculationType.SecondCorrectionSettlement));
-                calculationTypes.Add(new Option<CalculationType>(CalculationType.ThirdCorrectionSettlement));
-            }
-
-            return calculationTypes;
-        }
-
-        public IEnumerable<Option<MeteringPointType>> GetMeteringPointTypes()
-        {
-            var meteringPointTypes = new List<Option<MeteringPointType>>
-            {
-                new Option<MeteringPointType>(MeteringPointType.FlexConsumption),
-                new Option<MeteringPointType>(MeteringPointType.NonProfiledConsumption),
-                new Option<MeteringPointType>(MeteringPointType.Production),
-            };
-
-            if (marketRole != EicFunction.BalanceResponsibleParty && marketRole != EicFunction.EnergySupplier)
-            {
-                meteringPointTypes.Add(new Option<MeteringPointType>(MeteringPointType.Exchange));
-                meteringPointTypes.Add(new Option<MeteringPointType>(MeteringPointType.TotalConsumption));
-            }
-
-            return meteringPointTypes;
-        }
-
-        public bool GetIsGridAreaRequired() => marketRole == EicFunction.GridAccessProvider;
     }
 
     [OneOf]
